Override lensFetchInjectable in ApplicationBuilder - So that we can request from the backend routes
