+++ package.json
@@ -411,6 +411,7 @@
     "monaco-editor-webpack-plugin": "^5.0.0",
     "node-gyp": "^8.3.0",
     "node-loader": "^2.0.0",
+    "node-mocks-http": "^1.12.1",
     "nodemon": "^2.0.20",
     "playwright": "^1.29.2",
     "postcss": "^8.4.21",
