+++ src/main/router/create-handler-for-route.injectable.ts
@@ -5,8 +5,6 @@
 import { getInjectable } from "@ogre-tools/injectable";
 import type { ServerResponse } from "http";
 import loggerInjectable from "../../common/logger.injectable";
-import { lensAuthenticationHeader } from "../../common/vars/auth-header";
-import authHeaderValueInjectable from "../lens-proxy/auth-header-value.injectable";
 import type { LensApiRequest, Route } from "./route";
 import { contentTypes } from "./router-content-types";
 import { writeServerResponseFor } from "./write-server-response";
@@ -18,23 +16,9 @@ const createHandlerForRouteInjectable = getInjectable({
   id: "create-handler-for-route",
   instantiate: (di): CreateHandlerForRoute => {
     const logger = di.inject(loggerInjectable);
-    const authHeaderValue = `Bearer ${di.inject(authHeaderValueInjectable)}`;
 
     return (route) => async (request, response) => {
       const writeServerResponse = writeServerResponseFor(response);
-
-      if (route.requireAuthentication) {
-        const authHeader = request.getHeader(lensAuthenticationHeader);
-
-        if (authHeader !== authHeaderValue) {
-          writeServerResponse(contentTypes.txt.resultMapper({
-            statusCode: 401,
-            response: "Missing authorization",
-          }));
-
-          return;
-        }
-      }
 
       try {
         const result = await route.handler(request);
