+++ src/main/start-main-application/runnables/setup-lens-proxy.injectable.ts
@@ -6,15 +6,12 @@ import { getInjectable } from "@ogre-tools/injectable";
 import exitAppInjectable from "../../electron-app/features/exit-app.injectable";
 import lensProxyInjectable from "../../lens-proxy/lens-proxy.injectable";
 import loggerInjectable from "../../../common/logger.injectable";
-import lensProxyPortInjectable from "../../lens-proxy/lens-proxy-port.injectable";
 import isWindowsInjectable from "../../../common/vars/is-windows.injectable";
 import showErrorPopupInjectable from "../../electron-app/features/show-error-popup.injectable";
 import { beforeApplicationIsLoadingInjectionToken } from "../runnable-tokens/before-application-is-loading-injection-token";
 import buildVersionInjectable from "../../vars/build-version/build-version.injectable";
 import initializeBuildVersionInjectable from "../../vars/build-version/init.injectable";
-import lensProxyCertificateInjectable from "../../../common/certificate/lens-proxy-certificate.injectable";
-import fetchInjectable from "../../../common/fetch/fetch.injectable";
-import { Agent } from "https";
+import lensFetchInjectable from "../../../common/fetch/lens-fetch.injectable";
 
 const setupLensProxyInjectable = getInjectable({
   id: "setup-lens-proxy",
@@ -23,12 +20,10 @@ const setupLensProxyInjectable = getInjectable({
     const lensProxy = di.inject(lensProxyInjectable);
     const exitApp = di.inject(exitAppInjectable);
     const logger = di.inject(loggerInjectable);
-    const lensProxyPort = di.inject(lensProxyPortInjectable);
     const isWindows = di.inject(isWindowsInjectable);
     const showErrorPopup = di.inject(showErrorPopupInjectable);
     const buildVersion = di.inject(buildVersionInjectable);
-    const lensProxyCertificate = di.inject(lensProxyCertificateInjectable);
-    const fetch = di.inject(fetchInjectable);
+    const lensFetch = di.inject(lensFetchInjectable);
 
     return {
       id: "setup-lens-proxy",
@@ -45,11 +40,7 @@ const setupLensProxyInjectable = getInjectable({
         // test proxy connection
         try {
           logger.info("🔎 Testing LensProxy connection ...");
-          const versionResponse = await fetch(`https://127.0.0.1:${lensProxyPort.get()}/version`, {
-            agent: new Agent({
-              ca: lensProxyCertificate.get()?.cert,
-            }),
-          });
+          const versionResponse = await lensFetch("/version");
 
           const { version: versionFromProxy } = await versionResponse.json() as { version: string };
 
@@ -81,9 +72,6 @@ const setupLensProxyInjectable = getInjectable({
       runAfter: di.inject(initializeBuildVersionInjectable),
     };
   },
-
-  causesSideEffects: true,
-
   injectionToken: beforeApplicationIsLoadingInjectionToken,
 });
