+++ src/renderer/components/test-utils/get-application-builder.tsx
@@ -69,6 +69,11 @@ import fsInjectable from "../../../common/fs/fs.injectable";
 import joinPathsInjectable from "../../../common/path/join-paths.injectable";
 import homeDirectoryPathInjectable from "../../../common/os/home-directory-path.injectable";
 import { testUsingFakeTime } from "../../../common/test-utils/use-fake-time";
+import lensFetchInjectable from "../../../common/fetch/lens-fetch.injectable";
+import handleLensRequestInjectable from "../../../main/lens-proxy/handle-lens-request.injectable";
+import httpMocks from "node-mocks-http";
+import nodeFetchModuleInjectable from "../../../common/fetch/fetch-module.injectable";
+import stream from "stream";
 
 type Callback = (di: DiContainer) => void | Promise<void>;
 
@@ -212,6 +217,35 @@ export const getApplicationBuilder = () => {
       trayMenuIconPath = path;
     },
   }));
+
+  mainDi.override(lensFetchInjectable, (di) => {
+    return async (pathnameAndQuery, init) => {
+      const handleLensRequest = di.inject(handleLensRequestInjectable);
+      const { Headers, Response } = di.inject(nodeFetchModuleInjectable);
+
+      const url = new URL(pathnameAndQuery, "https://127.0.0.1");
+      const req = httpMocks.createRequest({
+        method: (init?.method ?? "get").toUpperCase() as any,
+        url: url.pathname,
+        params: url.searchParams,
+        body: (init?.body ?? undefined) as any,
+        headers: new Headers(init?.headers) as any,
+      });
+      const duplex = new stream.Duplex();
+      const res = httpMocks.createResponse({
+        req,
+        writableStream: duplex,
+      });
+
+      await handleLensRequest.handle(req, res);
+
+      return new Response(duplex, {
+        headers: new Headers(res._getHeaders() as Record<string, string>),
+        status: res._getStatusCode(),
+        statusText: res._getStatusMessage(),
+      });
+    };
+  });
 
   const allowedResourcesState = observable.set<string>();
